cast expected to match dtype float32

+++ tests/test_ml-notebook.py
@@ -49,7 +49,7 @@ def test_jax_random_number_generator():
     with jax.default_device(jax.devices("cpu")[0]):
         key = random.key(seed=42)
         x = random.normal(key=key)
-        np.testing.assert_allclose(x, -0.028305)
+        np.testing.assert_allclose(x, np.float32(-0.028305))
 
     # Test running on GPU (need to run locally)
     try:
